Strategy fixed about the headers.
Drop the usage of xfs.h and xfs_linux.h. Use xfsd.h instead. Anytime
when you need types like __s64 then include xfsd_types.h.

This method is only available when there is absolutely NO NEED for the
macro __KERNEL__ to be defined.

Besides, if there is any need for xfs.h to be included, it must be
changed.

If these requirments can't be fulfilled, xfs_types.h must be modified,
 too.